requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymysql==1.1.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
